+++ src/L10NSharp.Windows.Forms/UIComponents/LanguageChoosingDialog.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Drawing;
+using System.Threading.Tasks;
 using System.Windows.Forms;
 using L10NSharp.Windows.Forms.Translators;
 
@@ -20,10 +21,38 @@ namespace L10NSharp.Windows.Forms.UIComponents
 		void Application_Idle(object sender, EventArgs e)
 		{
 			Application.Idle -= Application_Idle;
-			_model.TranslateStrings(new BingTranslator("en", _model.RequestedCultureTwoLetterISOLanguageName));
-			_messageLabel.Text = _model.Message;
-			_OKButton.Text = _model.AcceptButtonText;
-			Text = _model.WindowTitle;
+			var targetCultureId = _model.RequestedCultureTwoLetterISOLanguageName;
+			TranslatorBase translator;
+			if (MicrosoftTranslator.IsConfigured)
+				translator = new MicrosoftTranslator("en", targetCultureId);
+			else
+				translator = new MyMemoryTranslator("en", targetCultureId);
+
+			// Translation makes a blocking network call (see TranslatorBase.TranslateText). Run it
+			// on a background thread so a slow or unresponsive endpoint can't freeze this dialog;
+			// only the (fast) UI update needs to happen on the UI thread, once translation is done.
+			Task.Run(() =>
+			{
+				_model.TranslateStrings(translator);
+				try
+				{
+					if (!IsDisposed)
+					{
+						BeginInvoke((Action)(() =>
+						{
+							if (IsDisposed)
+								return;
+							_messageLabel.Text = _model.Message;
+							_OKButton.Text = _model.AcceptButtonText;
+							Text = _model.WindowTitle;
+						}));
+					}
+				}
+				catch (ObjectDisposedException)
+				{
+					// Dialog was closed before translation finished; nothing to update.
+				}
+			});
 		}
 
 		public string SelectedLanguage;